pyproject.toml
[tool.poetry]
name = "stora"
version = "0.1.0"
description = "Firestore GUI Client"
authors = ["fbjorn"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9,<3.10"
google-cloud-firestore = "^2.1.1"
loguru = "^0.5.3"
pyinstaller = "^4.3"
invoke = "^1.5.0"
PySide6 = "^6.1.1"
Jinja2 = "^3.0.1"
qt-material = "^2.8.8"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
mypy = "^0.902"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
